+++ lib/auth/index.ts
@@ -7,7 +7,6 @@ import { DateTime } from 'luxon';
 import { TypedCache } from '@app/lib/cache.ts';
 import type { IUser, Permission } from '@app/lib/orm/index.ts';
 import { AccessTokenRepo, fetchPermission, fetchUserX } from '@app/lib/orm/index.ts';
-import * as orm from '@app/lib/orm/index.ts';
 import { intval } from '@app/lib/utils/index.ts';
 import NodeCache from '@app/vendor/node-cache.ts';
 
@@ -87,7 +86,8 @@ export async function byToken(accessToken: string): Promise<IAuth | null> {
   }
 
   const token = await AccessTokenRepo.findOne({
-    where: { accessToken, expires: orm.Gt(new Date()) },
+    accessToken,
+    expires: { $gt: new Date() },
   });
 
   if (!token) {
