+++ package.json
@@ -108,7 +108,7 @@
     "prettier": "^2.8.8",
     "prettier-plugin-jsdoc": "^0.4.2",
     "std-env": "^3.3.3",
-    "typescript": "~5.1.6",
+    "typescript": "~5.0.4",
     "vite": "^4.3.9",
     "vitest": "^0.32.2",
     "vitest-github-actions-reporter": "^0.10.0"
