+++ tsconfig.json
@@ -8,7 +8,6 @@
     "allowSyntheticDefaultImports": false,
     "useUnknownInCatchVariables": true,
     "esModuleInterop": false,
-    "emitDecoratorMetadata": true,
     "noImplicitAny": true,
     "noUncheckedIndexedAccess": true,
     "strict": true,
